+++ resources/views/profile/edit.blade.php
@@ -8,15 +8,19 @@
         <li class="nav-item">
             <a class="nav-link {{ $request->get('page') == "autentikasi" ? "active" : "" }}"  href="?page=autentikasi" >profil autentikasi</a>
         </li>
-        <li class="nav-item">
-            <a class="nav-link {{ $request->get('page') == "profil" ? "active" : "" }}" href="?page=profil">data profil</a>
-        </li>
+        @if( \Illuminate\Support\Facades\Auth::user()->tipe_akun != 2 || \Illuminate\Support\Facades\Auth::user()->id != 99)
+            <li class="nav-item">
+                <a class="nav-link {{ $request->get('page') == "profil" ? "active" : "" }}" href="?page=profil">data profil</a>
+            </li>
+        @endif
         <li class="nav-item">
             <a class="nav-link {{ $request->get('page') == "password" ? "active" : "" }}" href="?page=password">kata sandi</a>
         </li>
-        <li class="nav-item">
-            <a class="nav-link text-danger {{ $request->get('page') == "hapusakun" ? "active fw-bold" : "" }}" href="?page=hapusakun">hapus akun <i class="bi bi-exclamation-triangle"></i></a>
-        </li>
+        @if( \Illuminate\Support\Facades\Auth::user()->tipe_akun != 2 || \Illuminate\Support\Facades\Auth::user()->id != 99)
+            <li class="nav-item">
+                <a class="nav-link text-danger {{ $request->get('page') == "hapusakun" ? "active fw-bold" : "" }}" href="?page=hapusakun">hapus akun <i class="bi bi-exclamation-triangle"></i></a>
+            </li>
+        @endif
     </ul>
 
     <div class="py-12">
@@ -28,14 +32,36 @@
                             @include('profile.partials.update-profile-information-form')
                             @break
                         @case('profil')
-                            @include('profile.partials.update-users-entity-form', ['data' => $data])
+                                @if( \Illuminate\Support\Facades\Auth::user()->tipe_akun != 2 || \Illuminate\Support\Facades\Auth::user()->id != 99)
+                                    @include('profile.partials.update-users-entity-form', ['data' => $data])
+                                @else
+                                    <h2 class="text-lg font-medium text-gray-900">
+                                        {{ __('Selamat datang di halaman profil!') }}
+                                    </h2>
+
+                                    <p class="mt-1 text-sm text-gray-600">
+                                        {{ __("Pilih salah satu menu untuk update/melihat profil anda!") }}
+                                    </p>
+                                @endif
                             @break
                         @case('password')
                             @include('profile.partials.update-password-form')
                             @break
                         @case('hapusakun')
-                            @include('profile.partials.delete-user-form')
+
+                            @if( \Illuminate\Support\Facades\Auth::user()->tipe_akun != 2 || \Illuminate\Support\Facades\Auth::user()->id != 99)
+                                @include('profile.partials.delete-user-form')
+                            @else
+                                <h2 class="text-lg font-medium text-gray-900">
+                                    {{ __('Selamat datang di halaman profil!') }}
+                                </h2>
+
+                                <p class="mt-1 text-sm text-gray-600">
+                                    {{ __("Pilih salah satu menu untuk update/melihat profil anda!") }}
+                                </p>
+                            @endif
                             @break
+
                         @default
                             <h2 class="text-lg font-medium text-gray-900">
                                 {{ __('Selamat datang di halaman profil!') }}
